Moved heavy logic from Demographic constructor to FhirParserHelper class - updated action and unit tests - split parsing logic out into separated private methods
