+++ application/CohortManager/src/Functions/Shared/Common/Interfaces/IFhirParserHelper.cs
@@ -1,15 +1,21 @@
 namespace Common.Interfaces;
 
-using Microsoft.Azure.Functions.Worker.Http;
+using Hl7.Fhir.Model;
 using Model;
 
 public interface IFhirParserHelper
 {
+    /// <summary>
+    /// Parses FHIR JSON and converts it to a Demographic object
+    /// </summary>
+    /// <param name="json">The FHIR JSON string</param>
+    /// <returns>A Demographic model populated from FHIR data</returns>
+    Demographic ParseFhirJson(string json);
 
     /// <summary>
-    /// Parses FHIR response json and maps it to the Demographic model.
+    /// Maps a FHIR Patient object to a new Demographic object
     /// </summary>
-    /// <param name="json">Raw FHIR response json from PDS.</param>
-    /// <returns>Demographic</returns>
-    Demographic FhirParser(string json);
-};
+    /// <param name="patient">The FHIR Patient object</param>
+    /// <returns>A new Demographic model populated from FHIR data</returns>
+    Demographic MapPatientToDemographic(Patient patient);
+}
